services/rss-collector/requirements.txt

# AWS SDK
boto3==1.34.34
botocore==1.34.34

# RSS解析
feedparser==6.0.11
requests==2.31.0

# 数据库
psycopg2-binary==2.9.9
redis==5.0.1

# 工具库
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
schedule==1.2.1

# 日志
python-json-logger==2.0.7

# 文本处理（可选）
beautifulsoup4==4.12.3
lxml==5.1.0